requirements.txt
celery==4.1.0
flask-mongoengine==0.9.5
flask==0.12.2
gevent==1.2.2
google-cloud-storage==1.9.0
kombu==4.1.0
kubernetes==5.0.0
marshmallow-mongoengine==0.9.1
marshmallow==2.15.1
mongoengine==0.15.0
python-slugify==1.2.5
# dev
dotmap==1.2.20
mongomock==3.10.0
pytest==3.5.1
